Train_ticket_sales_service/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.6
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
billiard==3.6.4.0
bleach==5.0.1
boto3==1.26.21
botocore==1.29.21
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cssselect==1.1.0
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dicttoxml==1.7.4
dill==0.3.5.1
Django==4.1.3
django-celery==3.1.17
django-debug-toolbar==3.5.0
django-phonenumber-field==6.3.0
django-storages==1.13.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.19
fontawesomefree==6.2.0
fonttools==4.38.0
frozenlist==1.3.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geoip2==4.6.0
geopy==2.2.0
h11==0.13.0
h3==3.7.4
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
ip2geotools==0.1.6
IP2Location==8.8.1
isort==5.10.1
itypes==1.2.0
jaraco.classes==3.2.2
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.9.1
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
maxminddb==2.2.0
mccabe==0.7.0
more-itertools==8.14.0
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
pdfkit==1.0.0
phonenumbers==8.12.53
Pillow==9.2.0
pip-review==1.2.0
pkginfo==1.8.3
platformdirs==2.5.2
prompt-toolkit==3.0.33
pycodestyle==2.9.1
pycountry==22.3.5
pycparser==2.21
pydyf==0.5.0
Pygments==2.13.0
PyJWT==2.6.0
pylint==2.15.0
pyparsing==3.0.9
pyphen==0.13.0
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
ratelim==0.1.6
readme-renderer==37.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
s3transfer==0.6.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
timezonefinder==6.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
twine==4.0.1
typed-ast==1.5.4
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wkhtmltopdf==0.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
zipp==3.8.1
zopfli==0.2.1
